Made e310 appends specific to the sg1, tweaked idle
The idle fpga image built from source in 3.14.0 is
incompatible with UHD 3.14.0 (ad9361 block states
it expected 8, received 0).  This prevents the
device from deallocating without locking up the
E310, so we replace it with the default fpga image.
NOTE: the device will still lock up if you do not
deallocate before shutting down the node.

Signed-off-by: Thomas Goodwin <btgoodwin@geontech.com>